[build] Move cert registry into its own build target
