+++ gRPC-C++.podspec
@@ -626,8 +626,6 @@ Pod::Spec.new do |s|
                       'src/core/ext/upbdefs-generated/xds/type/matcher/v3/regex.upbdefs.h',
                       'src/core/ext/upbdefs-generated/xds/type/matcher/v3/string.upbdefs.h',
                       'src/core/ext/upbdefs-generated/xds/type/v3/typed_struct.upbdefs.h',
-                      'src/core/ext/xds/certificate_provider_factory.h',
-                      'src/core/ext/xds/certificate_provider_registry.h',
                       'src/core/ext/xds/certificate_provider_store.h',
                       'src/core/ext/xds/file_watcher_certificate_provider_factory.h',
                       'src/core/ext/xds/upb_utils.h',
@@ -860,6 +858,8 @@ Pod::Spec.new do |s|
                       'src/core/lib/security/authorization/grpc_server_authz_filter.h',
                       'src/core/lib/security/authorization/matchers.h',
                       'src/core/lib/security/authorization/rbac_policy.h',
+                      'src/core/lib/security/certificate_provider/certificate_provider_factory.h',
+                      'src/core/lib/security/certificate_provider/certificate_provider_registry.h',
                       'src/core/lib/security/context/security_context.h',
                       'src/core/lib/security/credentials/alts/alts_credentials.h',
                       'src/core/lib/security/credentials/alts/check_gcp_environment.h',
@@ -1486,8 +1486,6 @@ Pod::Spec.new do |s|
                               'src/core/ext/upbdefs-generated/xds/type/matcher/v3/regex.upbdefs.h',
                               'src/core/ext/upbdefs-generated/xds/type/matcher/v3/string.upbdefs.h',
                               'src/core/ext/upbdefs-generated/xds/type/v3/typed_struct.upbdefs.h',
-                              'src/core/ext/xds/certificate_provider_factory.h',
-                              'src/core/ext/xds/certificate_provider_registry.h',
                               'src/core/ext/xds/certificate_provider_store.h',
                               'src/core/ext/xds/file_watcher_certificate_provider_factory.h',
                               'src/core/ext/xds/upb_utils.h',
@@ -1720,6 +1718,8 @@ Pod::Spec.new do |s|
                               'src/core/lib/security/authorization/grpc_server_authz_filter.h',
                               'src/core/lib/security/authorization/matchers.h',
                               'src/core/lib/security/authorization/rbac_policy.h',
+                              'src/core/lib/security/certificate_provider/certificate_provider_factory.h',
+                              'src/core/lib/security/certificate_provider/certificate_provider_registry.h',
                               'src/core/lib/security/context/security_context.h',
                               'src/core/lib/security/credentials/alts/alts_credentials.h',
                               'src/core/lib/security/credentials/alts/check_gcp_environment.h',
