+++ gRPC-Core.podspec
@@ -947,9 +947,6 @@ Pod::Spec.new do |s|
                       'src/core/ext/upbdefs-generated/xds/type/matcher/v3/string.upbdefs.h',
                       'src/core/ext/upbdefs-generated/xds/type/v3/typed_struct.upbdefs.c',
                       'src/core/ext/upbdefs-generated/xds/type/v3/typed_struct.upbdefs.h',
-                      'src/core/ext/xds/certificate_provider_factory.h',
-                      'src/core/ext/xds/certificate_provider_registry.cc',
-                      'src/core/ext/xds/certificate_provider_registry.h',
                       'src/core/ext/xds/certificate_provider_store.cc',
                       'src/core/ext/xds/certificate_provider_store.h',
                       'src/core/ext/xds/file_watcher_certificate_provider_factory.cc',
@@ -1395,6 +1392,9 @@ Pod::Spec.new do |s|
                       'src/core/lib/security/authorization/matchers.h',
                       'src/core/lib/security/authorization/rbac_policy.cc',
                       'src/core/lib/security/authorization/rbac_policy.h',
+                      'src/core/lib/security/certificate_provider/certificate_provider_factory.h',
+                      'src/core/lib/security/certificate_provider/certificate_provider_registry.cc',
+                      'src/core/lib/security/certificate_provider/certificate_provider_registry.h',
                       'src/core/lib/security/context/security_context.cc',
                       'src/core/lib/security/context/security_context.h',
                       'src/core/lib/security/credentials/alts/alts_credentials.cc',
@@ -2111,8 +2111,6 @@ Pod::Spec.new do |s|
                               'src/core/ext/upbdefs-generated/xds/type/matcher/v3/regex.upbdefs.h',
                               'src/core/ext/upbdefs-generated/xds/type/matcher/v3/string.upbdefs.h',
                               'src/core/ext/upbdefs-generated/xds/type/v3/typed_struct.upbdefs.h',
-                              'src/core/ext/xds/certificate_provider_factory.h',
-                              'src/core/ext/xds/certificate_provider_registry.h',
                               'src/core/ext/xds/certificate_provider_store.h',
                               'src/core/ext/xds/file_watcher_certificate_provider_factory.h',
                               'src/core/ext/xds/upb_utils.h',
@@ -2345,6 +2343,8 @@ Pod::Spec.new do |s|
                               'src/core/lib/security/authorization/grpc_server_authz_filter.h',
                               'src/core/lib/security/authorization/matchers.h',
                               'src/core/lib/security/authorization/rbac_policy.h',
+                              'src/core/lib/security/certificate_provider/certificate_provider_factory.h',
+                              'src/core/lib/security/certificate_provider/certificate_provider_registry.h',
                               'src/core/lib/security/context/security_context.h',
                               'src/core/lib/security/credentials/alts/alts_credentials.h',
                               'src/core/lib/security/credentials/alts/check_gcp_environment.h',
