+++ grpc.gemspec
@@ -860,9 +860,6 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/upbdefs-generated/xds/type/matcher/v3/string.upbdefs.h )
   s.files += %w( src/core/ext/upbdefs-generated/xds/type/v3/typed_struct.upbdefs.c )
   s.files += %w( src/core/ext/upbdefs-generated/xds/type/v3/typed_struct.upbdefs.h )
-  s.files += %w( src/core/ext/xds/certificate_provider_factory.h )
-  s.files += %w( src/core/ext/xds/certificate_provider_registry.cc )
-  s.files += %w( src/core/ext/xds/certificate_provider_registry.h )
   s.files += %w( src/core/ext/xds/certificate_provider_store.cc )
   s.files += %w( src/core/ext/xds/certificate_provider_store.h )
   s.files += %w( src/core/ext/xds/file_watcher_certificate_provider_factory.cc )
@@ -1308,6 +1305,9 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/security/authorization/matchers.h )
   s.files += %w( src/core/lib/security/authorization/rbac_policy.cc )
   s.files += %w( src/core/lib/security/authorization/rbac_policy.h )
+  s.files += %w( src/core/lib/security/certificate_provider/certificate_provider_factory.h )
+  s.files += %w( src/core/lib/security/certificate_provider/certificate_provider_registry.cc )
+  s.files += %w( src/core/lib/security/certificate_provider/certificate_provider_registry.h )
   s.files += %w( src/core/lib/security/context/security_context.cc )
   s.files += %w( src/core/lib/security/context/security_context.h )
   s.files += %w( src/core/lib/security/credentials/alts/alts_credentials.cc )
