+++ src/python/grpcio/grpc_core_dependencies.py
@@ -398,7 +398,6 @@ CORE_SOURCE_FILES = [
     'src/core/ext/upbdefs-generated/xds/type/matcher/v3/regex.upbdefs.c',
     'src/core/ext/upbdefs-generated/xds/type/matcher/v3/string.upbdefs.c',
     'src/core/ext/upbdefs-generated/xds/type/v3/typed_struct.upbdefs.c',
-    'src/core/ext/xds/certificate_provider_registry.cc',
     'src/core/ext/xds/certificate_provider_store.cc',
     'src/core/ext/xds/file_watcher_certificate_provider_factory.cc',
     'src/core/ext/xds/xds_api.cc',
@@ -612,6 +611,7 @@ CORE_SOURCE_FILES = [
     'src/core/lib/security/authorization/grpc_server_authz_filter.cc',
     'src/core/lib/security/authorization/matchers.cc',
     'src/core/lib/security/authorization/rbac_policy.cc',
+    'src/core/lib/security/certificate_provider/certificate_provider_registry.cc',
     'src/core/lib/security/context/security_context.cc',
     'src/core/lib/security/credentials/alts/alts_credentials.cc',
     'src/core/lib/security/credentials/alts/check_gcp_environment.cc',
