+++ test/core/xds/xds_bootstrap_test.cc
@@ -20,11 +20,11 @@
 #include "absl/strings/numbers.h"
 #include "absl/strings/str_format.h"
 
-#include "src/core/ext/xds/certificate_provider_registry.h"
 #include "src/core/ext/xds/xds_bootstrap_grpc.h"
 #include "src/core/ext/xds/xds_client_grpc.h"
 #include "src/core/lib/gpr/env.h"
 #include "src/core/lib/gpr/tmpfile.h"
+#include "src/core/lib/security/certificate_provider/certificate_provider_registry.h"
 #include "test/core/util/test_config.h"
 
 namespace grpc_core {
